+++ src/renderer/components/markdown_content.tsx
@@ -2,7 +2,7 @@ import styled from "@emotion/styled";
 import ReactMarkdown from "react-markdown";
 
 import { ExternalLink as A } from "@/components/external_link";
-import { withFont } from "@/styles/withFont";
+import { withFont } from "@/styles/with_font";
 
 export const MarkdownContent = ({ content, className }: { className?: string; content: string }) => {
   return (
