+++ src/renderer/pages/quick_start/quick_start.tsx
@@ -7,7 +7,7 @@ import { useNavigate } from "react-router-dom";
 import { StepperDots } from "@/components/stepper_dots";
 import { useMousetrap } from "@/lib/hooks/use_mousetrap";
 import { QuickStartStep } from "@/lib/hooks/use_quick_start";
-import { platformTitleBarStyles } from "@/styles/platformTitleBarStyles";
+import { platformTitleBarStyles } from "@/styles/platform_title_bar_styles";
 
 import { AcceptRulesStep } from "./steps/accept_rules_step";
 import { ActivateOnlineStep } from "./steps/activate_online_step";
