+++ src/renderer/styles/has_border.ts
