+++ src/renderer/styles/platform_title_bar_styles.ts
