+++ src/renderer/styles/with_slippi_background.ts
